Refactoring mmap.
* Moved methods to package y.
* Added an Madvise method to control readahead setting.
