+++ y/mmap_windows.go
@@ -16,7 +16,7 @@
 
 // +build windows
 
-package table
+package y
 
 import (
 	"os"
@@ -24,15 +24,22 @@ import (
 	"unsafe"
 )
 
-func mmap(fd *os.File, size int64) ([]byte, error) {
+func Mmap(fd *os.File, write bool, size int64) ([]byte, error) {
+	protect := syscall.PAGE_READONLY
+	access := syscall.FILE_MAP_READ
+
+	if write {
+		protect = syscall.PAGE_READWRITE
+		access = syscall.FILE_MAP_WRITE
+	}
 	handler, err := syscall.CreateFileMapping(syscall.Handle(fd.Fd()), nil,
-		syscall.PAGE_READONLY, uint32(size>>32), uint32(size), nil)
+		uint32(protect), uint32(size>>32), uint32(size), nil)
 	if err != nil {
 		return nil, err
 	}
 	defer syscall.CloseHandle(handler)
 
-	mapData, err := syscall.MapViewOfFile(handler, syscall.FILE_MAP_READ, 0, 0, 0)
+	mapData, err := syscall.MapViewOfFile(handler, uint32(access), 0, 0, 0)
 	if err != nil {
 		return nil, err
 	}
@@ -41,6 +48,11 @@ func mmap(fd *os.File, size int64) ([]byte, error) {
 	return data, nil
 }
 
-func munmap(b []byte) error {
+func Munmap(b []byte) error {
 	return syscall.UnmapViewOfFile(uintptr(unsafe.Pointer(&b[0])))
+}
+
+func Madvise(b []byte, readahead bool) error {
+	// Do Nothing. We don’t care about this setting on Windows
+	return nil
 }
